requirements.txt
torch>=1.9.0
torchvision>=0.10.0
ultralytics>=8.0.0
opencv-python>=4.5.0
Pillow>=8.0.0
numpy>=1.21.0
matplotlib>=3.3.0
seaborn>=0.11.0
pandas>=1.3.0
PyYAML>=5.4.0
tqdm>=4.62.0
requests>=2.25.0
xmltodict>=0.12.0
albumentations>=1.0.0 